% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_cpp}
\alias{gibbs_cpp}
\title{generate gibbs using Rcpp}
\usage{
gibbs_cpp(N, mu1, mu2, sigma1, sigma2, rho)
}
\arguments{
\item{N}{the number of samples}

\item{mu1}{Mean value of the first parameter}

\item{mu2}{Mean value of the second parameter}

\item{sigma1}{Standard deviation of the first parameter}

\item{sigma2}{Standard deviation of the second parameter}

\item{rho}{correlation coefficient of the two parameter}
}
\value{
a random sample of size \code{n}
}
\description{
generate gibbs using Rcpp
}
